Prevent injection of custom nodes into the ast from the script tag
